Create RankAll
This function takes two arguments: an outcome name (outcome) and a hospital ranking (num). The function reads the outcome-of-care-measures.csv file and returns a 2-column data frame containing the hospital in each state that has the ranking specified in num. For example the function call
rankall("heart attack", "best") 
would return a data frame containing the names of the hospitals that are the best in their respective states for 30-day heart attack death rates.